+++ package.json
@@ -27,7 +27,7 @@
     "lint": "eslint \"src/**/*.ts\"",
     "format": "eslint \"src/**/*.ts\" --fix",
     "test": "jest --coverage",
-    "coverage": "jest --collectCoverageFrom=src/**.ts --coverage src",
+    "coverage": "jest --coverage",
     "watch": "tsc --watch"
   },
   "husky": {
@@ -83,10 +83,21 @@
   "jest": {
     "testEnvironment": "node",
     "roots": [
-      "<rootDir>/src"
+      "<rootDir>/src",
+      "<rootDir>/tests"
     ],
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    }
+    "preset": "ts-jest",
+    "coveragePathIgnorePatterns": [
+      "/node_modules/",
+      "/dist/",
+      "/tests/",
+      "/docs/",
+      "/data/",
+      "/config/",
+      "/@types/",
+      ".vscode",
+      ".github",
+      ".devcontainer"
+    ]
   }
 }
