+++ tests/regex.spec.ts
@@ -1,6 +1,6 @@
 /* eslint-disable jest/no-disabled-tests */
 // TODO: Fix commented tests
-import MDRegex from './regex';
+import MDRegex from '../src/util/regex';
 
 /** Tests the given regular expression
  *
