+++ tests/telegram.spec.ts
@@ -1,4 +1,4 @@
-import TelegramBot from './telegram';
+import TelegramBot from '../src/bots/telegram';
 
 jest.mock('request-promise-native');
